+++ apps/backend/src/config/dependencyConfigDefault.ts
@@ -41,7 +41,7 @@ import {
 } from '../eventHandlers/messageBroker';
 import { createApplicationEventBus } from '../events';
 import { ApplicationEvent } from '../events/applicationEvents';
-import { EssDownloadService } from '../factory/EssDownloadService';
+import { DefaultDownloadService } from '../factory/DefaultDownloadService';
 import { FapDataColumns } from '../factory/xlsx/FapDataColumns';
 import { getDataRow, populateRow } from '../factory/xlsx/FapDataRow';
 import { SkipAssetRegistrar } from '../services/assetRegistrar/skip/SkipAssetRegistrar';
@@ -104,4 +104,4 @@ mapValue(Tokens.ListenToMessageQueue, createSkipListeningHandler());
 mapValue(Tokens.ConfigureEnvironment, configureBaseEnvironment);
 mapValue(Tokens.ConfigureLogger, () => setLogger(new ConsoleLogger()));
 
-mapClass(Tokens.DownloadService, EssDownloadService);
+mapClass(Tokens.DownloadService, DefaultDownloadService);
