+++ apps/backend/src/config/dependencyConfigESS.ts
@@ -37,7 +37,7 @@ import {
   createPostToRabbitMQHandler,
 } from '../eventHandlers/messageBroker';
 import { createApplicationEventBus } from '../events';
-import { EssDownloadService } from '../factory/EssDownloadService';
+import { DefaultDownloadService } from '../factory/DefaultDownloadService';
 import { FapDataColumns } from '../factory/xlsx/FapDataColumns';
 import { getDataRow, populateRow } from '../factory/xlsx/FapDataRow';
 import { EAMAssetRegistrar } from '../services/assetRegistrar/eam/EAMAssetRegistrar';
@@ -101,4 +101,4 @@ mapValue(
 );
 mapValue(Tokens.ConfigureLogger, configureGraylogLogger);
 
-mapClass(Tokens.DownloadService, EssDownloadService);
+mapClass(Tokens.DownloadService, DefaultDownloadService);
