+++ apps/backend/src/datasources/FileDataSource.ts
@@ -1,7 +1,7 @@
 import { ReadStream } from 'pg-large-object';
 
 import { FileMetadata } from '../models/Blob';
-import { FilesMetadataFilter } from './../resolvers/queries/FilesMetadataQuery';
+import { FilesMetadataFilter } from '../resolvers/queries/FilesMetadataQuery';
 export interface FileDataSource {
   prepare(fileId: string, output: string): Promise<string>;
   // Read
