+++ apps/backend/src/datasources/mockups/FileDataSource.ts
@@ -1,7 +1,7 @@
 import { ReadStream } from 'pg-large-object';
 
 import { FileMetadata } from '../../models/Blob';
-import { FileDataSource } from '../IFileDataSource';
+import { FileDataSource } from '../FileDataSource';
 import { FilesMetadataFilter } from './../../resolvers/queries/FilesMetadataQuery';
 
 export default class FileDataSourceMock implements FileDataSource {
