+++ apps/backend/src/factory/StfcDownloadService.ts
@@ -1,14 +1,11 @@
 import querystring from 'querystring';
-import { Readable } from 'stream';
-import { ReadableStream } from 'stream/web';
 
-import { logger } from '@user-office-software/duo-logger';
 import contentDisposition from 'content-disposition';
 import { Request, Response, NextFunction } from 'express';
 import { container } from 'tsyringe';
 
 import { Tokens } from '../config/Tokens';
-import { FileDataSource } from '../datasources/IFileDataSource';
+import { FileDataSource } from '../datasources/FileDataSource';
 import { Role } from '../models/Role';
 import {
   MetaBase,
@@ -17,10 +14,9 @@ import {
   XLSXType,
   ZIPType,
   DownloadService,
+  fetchDataAndStreamResponse,
 } from './DownloadService';
 import { ProposalPDFData } from './pdf/proposal';
-
-const ENDPOINT = process.env.USER_OFFICE_FACTORY_ENDPOINT;
 
 export class StfcDownloadService implements DownloadService {
   async callFactoryService<TData, TMeta extends MetaBase>(
@@ -31,19 +27,17 @@ export class StfcDownloadService implements DownloadService {
     res: Response,
     next: NextFunction
   ) {
-    if (!ENDPOINT) {
-      logger.logError(
-        'Could not start application: the `USER_OFFICE_FACTORY_ENDPOINT` environment variable is missing. Exiting.',
-        {}
-      );
-      process.exit(1);
-    }
-
     let isPdfAvailable = false;
     let proposalPdfData = null;
-    const data = properties?.data[0] as ProposalPDFData;
+    let fileName = '';
 
-    if (properties?.data?.length === 1) {
+    if (
+      properties?.data?.length === 1 &&
+      type === PDFType.PROPOSAL &&
+      downloadType === DownloadType.PDF
+    ) {
+      const data = properties?.data[0] as ProposalPDFData;
+
       const fileDataSource = container.resolve<FileDataSource>(
         Tokens.FileDataSource
       );
@@ -53,15 +47,14 @@ export class StfcDownloadService implements DownloadService {
 
       if (proposalPdfData) {
         isPdfAvailable = true;
+        fileName = `${data.proposal.proposalId}_${
+          data.principalInvestigator.lastname
+        }_${data.proposal.created.getUTCFullYear()}.pdf`;
       }
     }
 
-    if (isPdfAvailable && proposalPdfData) {
+    if (isPdfAvailable && proposalPdfData && fileName != '') {
       try {
-        const fileName = `${data.proposal.proposalId}_${
-          data.principalInvestigator.lastname
-        }_${data.proposal.created.getUTCFullYear()}.pdf`;
-
         res.setHeader('Content-Disposition', contentDisposition(fileName));
         res.setHeader('x-download-filename', querystring.escape(fileName));
         res.setHeader('content-type', 'application/pdf');
@@ -76,74 +69,14 @@ export class StfcDownloadService implements DownloadService {
         });
       }
     } else {
-      const controller = new AbortController();
-
-      try {
-        let gotResponse = false;
-
-        req.on('close', () => {
-          if (!gotResponse) {
-            controller.abort();
-          }
-        });
-
-        req.on('end', () => {
-          gotResponse = true;
-        });
-
-        const factoryResp = await fetch(`${ENDPOINT}/${downloadType}/${type}`, {
-          signal: controller.signal,
-          headers: {
-            Accept: 'application/json',
-            'Content-Type': 'application/json',
-          },
-          method: 'POST',
-          body: JSON.stringify(properties),
-        });
-
-        if (!factoryResp.ok) {
-          throw new Error(await factoryResp.text());
-        }
-
-        const factoryRespBody = factoryResp.body;
-        if (!factoryRespBody) {
-          return;
-        }
-
-        const readableStream = Readable.fromWeb(
-          factoryRespBody as ReadableStream,
-          {
-            signal: controller.signal,
-          }
-        );
-
-        readableStream.on('error', (err) => {
-          next({
-            error: err.toString(),
-            message: `Could not download generated ${downloadType}/${type}`,
-          });
-        });
-
-        const filename =
-          properties.data.length > 1
-            ? properties.meta.collectionFilename
-            : properties.meta.singleFilename;
-
-        const contentTypeHeaders = factoryResp.headers.get('content-type');
-
-        res.setHeader('Content-Disposition', contentDisposition(filename));
-        res.setHeader('x-download-filename', querystring.escape(filename));
-        if (contentTypeHeaders) {
-          res.setHeader('content-type', contentTypeHeaders);
-        }
-
-        readableStream.pipe(res);
-      } catch (error) {
-        next({
-          error,
-          message: `Could not download generated ${downloadType}/${type}`,
-        });
-      }
+      fetchDataAndStreamResponse(
+        downloadType,
+        type,
+        properties,
+        req,
+        res,
+        next
+      );
     }
   }
 }
