+++ apps/backend/src/queries/FileQueries.ts
@@ -1,7 +1,7 @@
 import { inject, injectable } from 'tsyringe';
 
 import { Tokens } from '../config/Tokens';
-import { FileDataSource } from '../datasources/IFileDataSource';
+import { FileDataSource } from '../datasources/FileDataSource';
 import { FilesMetadataFilter } from './../resolvers/queries/FilesMetadataQuery';
 
 @injectable()
